+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/projects/ProjectsRendererExtension.java
@@ -11,15 +11,15 @@ import java.util.Optional;
 import static net.splitcells.dem.data.set.Sets.setOfUniques;
 
 public interface ProjectsRendererExtension {
-    Optional<RenderingResult> renderFile(String path, ProjectsRenderer projectsRenderer, Config config);
+    Optional<RenderingResult> renderFile(String path, ProjectsRendererI projectsRendererI, Config config);
 
-    default Perspective extendProjectLayout(Perspective layout, ProjectsRenderer projectsRenderer) {
+    default Perspective extendProjectLayout(Perspective layout, ProjectsRendererI projectsRendererI) {
         return layout;
     }
 
-    Set<Path> projectPaths(ProjectsRenderer projectsRenderer);
+    Set<Path> projectPaths(ProjectsRendererI projectsRendererI);
 
-    default Set<Path> relevantProjectPaths(ProjectsRenderer projectsRenderer) {
-        return projectPaths(projectsRenderer);
+    default Set<Path> relevantProjectPaths(ProjectsRendererI projectsRendererI) {
+        return projectPaths(projectsRendererI);
     }
 }